+++ lib/src/data/ble/link/ble_link_base.dart
@@ -7,16 +7,12 @@ import 'package:flutter/foundation.dart';
 import '../../../domain/transport/transport_link.dart';
 
 abstract base class BleLinkBase implements TransportLink {
-  BleLinkBase({
-    required this.appName,
-    required String serviceId,
-    required String characteristicId,
-  }) : serviceUuid = UUID.fromString(serviceId),
-       characteristicUuid = UUID.fromString(characteristicId);
+  BleLinkBase({required this.appName, required String serviceId, required String characteristicId})
+    : serviceUuid = UUID.fromString(serviceId),
+      characteristicUuid = UUID.fromString(characteristicId);
 
   @override
-  Stream<Uint8List> get incomingRawMessageStream =>
-      _incomingRawMessageController.stream;
+  Stream<Uint8List> get incomingRawMessageStream => _incomingRawMessageController.stream;
 
   @override
   Future<void> dispose() async {